requirements.txt
streamlit==1.1.0
streamlit_chat==0.0.2.2
torch
transformers==4.29.2
sentence_transformers
langchain
sacremoses
chromadb
pytrials
mysql-connector-python